flask/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.0.0
psycopg2-binary==2.9.5
python-dotenv==1.0.0
pandas
gunicorn

